pyproject.toml
[tool.poetry]
name = "fastapiproj"
version = "0.1.0"
description = ""
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["classic"], version = "^0.30.6"}
fastapi = "^0.112.1"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.8.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.10"
aio-pika = "^9.4.3"
loguru = "^0.7.2"
psycopg2-binary = "^2.9.9"
stripe = "^11.1.0"
httpx = "^0.27.2"
pytest = "^8.3.3"
pytz = "^2024.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
